Implement creator and admin fees
This commit introduces an admin address to the CreatorToken, and introduces
fees which are taken when a user buys a creator token. The fee percentages
are stored as hardcoded values in BIPS. The fees are *additive*, meaning the
price of the NFT is calculated, then the fees are added ON TOP of this to
get the total owed.