Implement the Creator Token Factory
Creates a CreatorTokenFactory with a deploy method that takes a configuration
struct and uses it to deploy a bonding curve contract and creator token
contract that uses said bonding curve.

The factory is integrated into the pre-existing Showtime Verifier infrastructure.
It requires the deployment be attested to by a signature from a whitelisted
verifier address before it can be deployed. The factory contract has helper
methods used to encode and generate the digest that will be signed by the verifier.

The tests for the token factory fork the Base mainnet in order to interact with
the existing onchain verifier infrastructure.